+++ DiscoveryService/Messages/KademliaRequests/FindValueRequest.cs
@@ -48,7 +48,7 @@ namespace LUC.DiscoveryService.Messages.KademliaRequests
         {
             if (writer != null)
             {
-                writer.Write((UInt32)MessageOperation);
+                writer.Write((UInt16)MessageOperation);
                 writer.Write(RandomID.ToString());
                 writer.Write(Sender.ToString());
                 writer.Write(ContactId.ToString());
